refactor(workshop): normalize sprint 06 contracts and docs
