+++ .todo/epics/epic-workshop-architecture-refactor-2026-08-03/sprints/02-shared-route-contract-ownership.md
@@ -1,6 +1,6 @@
 # Sprint 02: Shared Route and Contract Ownership
 
-**Status:** Complete — implemented and verified on 2026-08-03
+**Status:** Complete — merged in PR #103
 
 **Branch:** `sprint/workshop-architecture-refactor-02-shared-ownership` -> `epic/workshop-architecture-refactor`
