+++ .todo/tech-debt/2026-07-25-workshop-god-files.md
@@ -10,26 +10,28 @@
 
 ## Problem
 
-Two files absorb every new Workshop feature instead of shedding any:
+The original debt was two files absorbing every new Workshop feature instead
+of shedding any. Phases 4 and 5 have reduced and narrowed both substantially,
+but Phase 7 still needs to certify that each remaining facade has one legible
+primary responsibility:
 
 | File | Lines (post-13A) | Nearest sibling |
 | --- | --- | --- |
-| `packages/core/src/application/services/workshop/WorkshopSessionService.ts` | 2,894 (PR #98 review remediation, 2026-07-31) | `RunWorkshopToolSidePass.ts` (305) |
-| `packages/core/src/application/handlers/domain/WorkshopHandler.ts` | 2,959 (PR #98, 2026-07-31) | `WorkshopSessionMessageHandler.ts` (303) |
+| `packages/core/src/application/services/workshop/WorkshopSessionService.ts` | 2,121 (Phase 5, 2026-08-06) | `session/WorkshopPassageScope.ts`, `WorkshopParticipantRoster.ts`, `WorkshopTodoLedger.ts`, `WorkshopTurnLedger.ts` |
+| `packages/core/src/application/handlers/domain/workshop/WorkshopHandler.ts` | 1,653 (Phase 4, 2026-08-06) | Five shared route owners plus three widget route owners |
 
-`CLAUDE.md`'s anti-pattern checklist flags any file over 500 lines. Every other
-file in both directories is comfortably under 1,000. Both files were already
-past the threshold before Sprint 13A, so this is a worsening trend line rather
-than a regression introduced by that sprint — which is why the review deferred
-it rather than blocking on it.
+Line count remains pressure evidence, not the completion rule. The open question
+is whether the remaining lines form one coherent aggregate facade and one
+coherent room/run orchestrator, not whether either crosses an arbitrary limit.
 
 ## Why it matters
 
-The codebase has already proved the fix on these exact files. `WorkshopHandler`
-shed its session-lifecycle routes into `WorkshopSessionMessageHandler`, wired
-through `registerRoutes(router, registerMutation)`; tool runs came out of the
-aggregate as `RunWorkshopToolSidePass`. Both decompositions shipped without
-incident. The pattern exists and works; new work simply isn't using it.
+The codebase has now applied the proven pattern across both files.
+`WorkshopHandler` composes named context, excerpt/scope, session, todo,
+standing-directive, and widget route owners. `WorkshopSessionService` delegates
+to named passage, participant, todo, turn, widget-config, standing-directive,
+persistence, and time collaborators while remaining the sole whole-session
+mutation boundary.
 
 The cost is concentration risk: a single file is becoming the place every
 Workshop change has to touch, which makes review harder, merge conflicts more
@@ -49,36 +51,35 @@ lines. The two larger files below remain the tracked debt.
   `WorkshopSessionService` keeps the aggregate-facing methods but no longer
   stores that collection/counter or its Gesture clone/summary rules inline.
   This closes PR #96 finding 12 while leaving the broader god-file target open.
-- **2026-07-31 — scope/context IPC extraction sequenced as optional Sprint
-  02C.** After the Lexical Gravity follow-up (02B-A), reassess whether the eight-route cluster
-  remains a pure-move seam. If so, extract it before Prose Controller; if the
-  move requires behavior or run/send-pipeline changes, skip it and leave this
-  debt open. See the [Sprint 02C plan](../epics/epic-conversation-widgets-2026-07-22/sprints/02c-workshop-scope-context-ipc-extraction.md).
+- **2026-08-05 — scope/context IPC extraction completed in Phase 4.**
+  `WorkshopContextHandler` and `WorkshopExcerptScopeHandler` now own the
+  measured route families; the earlier optional Sprint 02C seam is superseded.
 - **2026-07-31 — standing-directive presentation extracted during PR #98
   review.** Widget-local frame building, summaries, marker copy, and display
   formatting moved out of `WorkshopSessionService`; the aggregate now delegates
   through the standing presentation registry. This reduced the session file
   from the reviewed ~2,950 lines to 2,894 while leaving the broader target open.
+- **2026-08-06 — aggregate state machines extracted in Phase 5.** Passage
+  scope, participant roster, todo, turn, persistence, and time responsibilities
+  now have named owners behind `WorkshopSessionService`, reducing it to 2,121
+  lines without exposing an internal ledger to handlers.
+- **2026-08-06 — Sprint 06 deliberately leaves this record open.** Okey accepted
+  D3: the `WorkshopHandler` → `WorkshopRoomHandler` naming tension and final
+  facade-cohesion verdict belong to Phase 7, not the contract-normalization diff.
 
-## Candidate seams
+## Remaining closure questions
 
-Named by the reviewers, in rough order of cohesion:
+1. Does `WorkshopHandler`'s remaining room/run orchestration justify its name,
+   or would `WorkshopRoomHandler` make the boundary materially easier to find?
+2. Are `WorkshopHandler` and `WorkshopSessionService` now narrow facades over
+   their named collaborators under representative end-to-end traces?
+3. Does any remaining responsibility have an independent invariant and reason
+   to change, rather than merely contributing line count? If not, stop
+   extracting.
 
-1. **Scope/shelf state machine** (aggregate) — `setSessionScope`,
-   `repinShelvedExcerpt`, `adoptShelvedExcerpt`, `scopeTransition`,
-   `recordScopeChange`, `hostDeliveredExcerptVersion`, `excerptDeliveryReason`.
-   Cohesive and independently testable; `WorkshopSessionScope.test.ts` already
-   tests it as a unit.
-2. **Session-scope IPC routes** (handler) — `WORKSHOP_SET_SESSION_SCOPE`,
-   `WORKSHOP_REPIN_EXCERPT`, and the context-attachment trio. The PR's own doc
-   comments already group these under a *"Session scope (Sprint 13A §2/§4)"*
-   banner, which is the seam asking to be cut. Follows
-   `WorkshopSessionMessageHandler` exactly. Sequenced as optional Sprint 02C;
-   it remains non-blocking for Prose Controller.
-3. **Webview scope-transition callbacks** (`WorkshopApp.tsx`) —
-   `startOpenConversation`, `continueWithExcerpt`, `hasWorkingSet`, and the
-   session-confirm branches would fit a `useWorkshopSessionBoundary` hook,
-   following the Tripartite Hook pattern the sibling domain hooks use.
+Presentation-modal workflow ownership is tracked separately in
+`2026-08-04-workshop-modal-workflow-ownership.md`; it is not a reason to reopen
+the aggregate or handler boundaries here.
 
 ## Completion criteria
