+++ packages/core/src/application/services/workshop/widgets/lexicalGravity/LexicalGravityConfigCodec.ts
@@ -21,14 +21,31 @@ import {
   shapeError
 } from '@/application/services/workshop/persistedValidation';
 import {
-  isLexicalGravityReach,
-  isLexicalGravityWeight,
-  LEXICAL_GRAVITY_REACH,
-  LEXICAL_GRAVITY_WEIGHT
-} from '@shared/constants/workshopWidgets';
-import {
   assertLexicalGravityLensRenderable
 } from './LexicalGravityDirective';
+
+export const LEXICAL_GRAVITY_WEIGHT = Object.freeze({
+  minimum: 10,
+  maximum: 100,
+  step: 5
+});
+
+export const LEXICAL_GRAVITY_REACH = Object.freeze({
+  minimum: 1,
+  maximum: 3,
+  values: [1, 2, 3] as const
+});
+
+export function isLexicalGravityWeight(value: unknown): value is number {
+  return Number.isSafeInteger(value)
+    && (value as number) >= LEXICAL_GRAVITY_WEIGHT.minimum
+    && (value as number) <= LEXICAL_GRAVITY_WEIGHT.maximum
+    && (value as number) % LEXICAL_GRAVITY_WEIGHT.step === 0;
+}
+
+export function isLexicalGravityReach(value: unknown): value is 1 | 2 | 3 {
+  return LEXICAL_GRAVITY_REACH.values.some((candidate) => candidate === value);
+}
 
 const BUDGET = PROMPT_BUDGETS.workshopWidgets;
 const SLUG = /^[a-z0-9]+(?:-[a-z0-9]+)*$/;
